renamed dashboard to AuthChecker and will make a new Dashboard component render within it
