+++ src/components/Admin/AuthChecker.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import { AuthConsumer } from './AuthContext';
+
+const AuthChecker = props => {
+    return(
+        <AuthConsumer>
+            { value  => (
+                value.isAuthenticated 
+                ?
+                    <div>This will be the Dashboard componenet</div>
+                    // <Dashboard /> will be loaded here
+                :
+                    <div>Not Logged In</div>
+            )}
+        </AuthConsumer>
+    );
+}
+
+export default AuthChecker;
