+++ src/components/Admin/Dashboard.jsx
@@ -1,19 +1,4 @@
 import React from 'react';
 
-import { AuthConsumer } from '../elements/AuthContext';
+// this will be the actual componenet that is loaded after redirect to "Dashboard" after clearing AuthChecker
 
-const Dashboard = props => {
-    return(
-        <AuthConsumer>
-            { value  => (
-                value.isAuthenticated 
-                ?
-                    <div>This will be the Dashboard componenet</div>
-                :
-                    <div>Not Logged In</div>
-            )}
-        </AuthConsumer>
-    );
-}
-
-export default Dashboard;
