+++ src/sass/_signin.scss
@@ -96,8 +96,8 @@
         border-left: 1px solid #000;
         border-right: 1px solid #000;
         border-bottom: 1px solid #000;
-        border-top-left-radius: 5px;
-        border-top-right-radius: 5px;
+        border-top-left-radius: 4.4px;
+        border-top-right-radius: 4.4px;
     }
 
     input[type=password].signIn-password {
